+++ src/olmo_core/eval/__init__.py
@@ -1,0 +1,8 @@
+"""
+Metrics and evaluator classes.
+"""
+
+from .evaluator import Evaluator
+from .metrics import MeanMetric, Metric
+
+__all__ = ["Evaluator", "Metric", "MeanMetric"]