+++ src/olmo_core/train/callbacks/__init__.py
@@ -1,7 +1,8 @@
-from .callback import Callback
+from .callback import Callback, CallbackConfig
 from .checkpointer import CheckpointerCallback, CheckpointRemovalStrategy
 from .config_saver import ConfigSaverCallback
 from .console_logger import ConsoleLoggerCallback
+from .evaluator_callback import EvaluatorCallback, LMEvaluatorCallbackConfig
 from .garbage_collector import GarbageCollectorCallback
 from .gpu_memory_monitor import GPUMemoryMonitorCallback
 from .grad_clipper import GradClipperCallback
@@ -13,10 +14,13 @@ from .wandb import WandBCallback
 
 __all__ = [
     "Callback",
+    "CallbackConfig",
     "CheckpointerCallback",
     "CheckpointRemovalStrategy",
     "ConfigSaverCallback",
     "ConsoleLoggerCallback",
+    "EvaluatorCallback",
+    "LMEvaluatorCallbackConfig",
     "GarbageCollectorCallback",
     "GPUMemoryMonitorCallback",
     "GradClipperCallback",